Develop

+++ src/controllers/UserController.ts
@@ -1,6 +1,7 @@
 import { response, Response } from "express";
-import { Body, Delete, Get, JsonController, OnUndefined, Param, Post, QueryParams, Res } from "routing-controllers";
+import { Body, Delete, Get, JsonController, OnUndefined, Param, Patch, Post, QueryParams, Res } from "routing-controllers";
 import { UserEntity } from "../domains/entities/UserEntity";
+import { UserModel } from "../domains/models/UserModel";
 import { UserType } from "../domains/models/UserType";
 import IUserRepository from "../domains/repositories/UserRepository";
 import UserService from "../domains/services/UserService";
@@ -10,40 +11,46 @@ import { UserName } from "../domains/valueobjects/UserName";
 import UserSQLite from "../infrastructures/sqlite/UserSQLite";
 import { UserDto } from './dto/UserDto';
 
-@JsonController()
+@JsonController('/user')
 export class UserController {
     private userService: UserService = UserService.create();
     private userRepository: IUserRepository = UserSQLite.create();
 
     /**
-     * ユーザの全取得
-     * @param query 
-     * @returns 
+     * ユーザの全取得 
+     * @returns {Array<UserEntity>} ユーザ情報配列
      */
-    @Get('/users')
+    @Get('/findAll')
     @OnUndefined(404)
     public async findAll() {
-        return await this.userRepository.findAll();
+        const datalist: Array<UserEntity> = await this.userRepository.findAll();
+        const result: Array<UserDto> = new Array<UserDto>;
+        for (const data of datalist) {
+            result.push(UserModel.create(data).responseBody());
+        }
+        return result;
     }
 
     /**
      * ユーザの取得
-     * @param query 
-     * @returns 
+     * @param {number} id 検索したいユーザID 
+     * @returns {Response}
      */
-    @Get('/users/:id')
+    @Get('/find/:id')
     @OnUndefined(404)
     public async find(@Param('id') id: number) {
-        return await this.userRepository.find(id);
+        console.log(`ID: ${id}`);
+        const user: UserEntity = await this.userRepository.find(id);
+        return UserModel.create(user).responseBody();
     }
 
     /**
      * ユーザの追加・更新
-     * @param body 
-     * @returns 
+     * @param {UserDto} body クライアント側から発行されたデータ
+     * @returns {Response}
      */
-    @Post('/users/add')
-    public async add(@Body() body: UserDto, @Res() response: Response) {
+    @Post('/insert')
+    public async insert(@Body() body: UserDto, @Res() response: Response) {
         const hashedPassword = this.userService.hashPassword(body.password);
         const entity: UserEntity = UserEntity.create(0, {
             username: UserName.create({ name: body.username }),
@@ -51,8 +58,8 @@ export class UserController {
             email: body.email,
             password: hashedPassword,
             imageUrl: body.imageUrl,
-            createdAt: Time.create({date: '' }),
-            updatedAt: Time.create({date: '' })
+            createdAt: Time.create({ value: new Date().toISOString() }),
+            updatedAt: Time.create({ value: new Date().toISOString() })
         })
         await this.userRepository.insert(entity);
         return response.send('OK');
@@ -60,10 +67,10 @@ export class UserController {
 
     /**
      * ユーザの追加・更新
-     * @param body 
-     * @returns 
+     * @param {UserDto} body クライアント側から発行されたデータ
+     * @returns {Response}
      */
-    @Post('/users/update/:id')
+    @Post('/update/:id')
     public async update(@Param('id') id: number, @Body() body: UserDto, @Res() response: Response) {
         const oldEntity: UserEntity = await this.userRepository.find(id);
         if (!this.userService.comparePassword(body.password, oldEntity.password)) {
@@ -75,20 +82,27 @@ export class UserController {
             email: body.email,
             password: body.password,
             imageUrl: body.imageUrl,
-            createdAt: Time.create({date: new Date().toLocaleString()}),
-            updatedAt: Time.create({date: new Date().toLocaleString()})
+            createdAt: Time.create({ value: oldEntity.createdAt }),
+            updatedAt: Time.create({ value: new Date().toISOString() })
         }));
         return response.send('更新が完了しました。');
     }
 
     /**
      * ユーザの削除
-     * @param id 
-     * @returns 
+     * @param {number} id ユーザID
+     * @returns {Response}
      */
-    @Delete('/users/:id')
-    public async delete(@Param('id') id: number) {
+    @Delete('/remove/:id')
+    public async remove(@Param('id') id: number) {
+        console.log(`ID: ${id}`);
+        const entity: UserEntity = await this.userRepository.find(id);
+        // 管理者以外の場合
+        if (entity.authority !== UserType.ADMIN) {
+            return response.send('権限がありません。');
+        }
+        // 削除処理の実行
         await this.userRepository.remove(id);
-        return response.send('OK');
+        return response.send('削除されました。');
     }
 }
